Fixed weapon struct
When refactoring tokenIDs to hashes I accidently changed the weapon id to a hash. Weapon id (uuid) and hash a two separate things. The hash is unique for every weapon owned by a user and returned in damage history. The id is unique per in-game weapon/model, checked when weapon stats are given by the server.