Removed DX11 config option
This should fix Unreal Editor compiling shaders every time you open it.

The settings are added/removed using the build batch script as they're still needed for the production build.
I tried just adding it to the Engine.ini file in the final build folder but had some problems. It seems like either the `Nanite.RequireDX12` and/or the `DefaultGraphicsRHI` setting need to be set before the build.