Removed RiderLink from Plugins folder
To avoid issues with different developers using different versions of Rider and RiderLink; I removed it from the project's plugins folder.

You can install RiderLink to the Engine but there still seems to be issues with that. Developers can still install it to the project directly but since it's now in .gitignore it won't be pushed to the repo.