Fixed shooting anim issue
Noticed when playing as a mech, not sure if it affected the AI at all. Sometimes clicking once would result in the mech shooting multiple times as the check to see if the animation finished failed sometimes. This was due to the animations being set to loop, so the == 0 check doesn't work reliably.