Battle Replay Tool + Additional Logging (#277)
* Started battle replay tool

Started working on a tool to replay a battle from a file that contains messages from the server (BATTLE:INIT and abilities)

* Continued battle replay

* Load battle replays from clipboard

* Fixed shooting anim issue

Noticed when playing as a mech, not sure if it affected the AI at all. Sometimes clicking once would result in the mech shooting multiple times as the check to see if the animation finished failed sometimes. This was due to the animations being set to loop, so the == 0 check doesn't work reliably.