Lowered UserActionSpawnPoint on Forest and Hive
Airstrike jets were spawning way too high on the forest map, causing airstrikes to be very delayed.
Lowered it on The Hive map as well as there are no buildings that would get in the way of the jets. Might as well get faster airstrike and landmine drops on it.